Adding logback support in opal-library
We decided to switch to the logback logging framework in order to improve
the logging facilities. To distinguish the types of logs (regardless the
level), logback introduces implements the notion of Marker introduced by
slf4j.
I created a new enum type to define the type of log: global, GP or
Application-level APDU, incoming APDU at the Global Platform (GP) or
application level, outgoing APDU at the GP or application level, and
command title. They are defined in the LogType enumeration as markers for
the logging framework. There is also a notion of belonging between the
markers.
Moreover, I needed to create a filter for logback (MarkerFilter) which
takes into account the notion of markers to distinguish the log
types. Indeed, the default implementation of such a matching filter on
marker names does not correspond to our implementation need (the "null"
marker must be a mismatch and not neutral).
Finally, I added the logback.xml file which contains the logging properties
with a new message layout.