Adding methods to CardConfigFactory and bug fixes
I added the methods from Yorick Lesecque in order to manage
effectively the card configurations from the XML files. It important
to notice that we no longer deal with the classloader to open
configuration files (it does not allow to write in files).
There are also some little bug fixes in the Conversion class by
Yorick Lesecque.
Finally, I corrected a little typo in a method name:
compudeAndVerifyRMac has been renamed to computeAndVerifyRMac in
the GP2xCommands class.