Remove opal-library from repository
I removed the opal-library folder from the repository, because it is no
longer needed: the modification we made on the library have been merged
in the official opal-library repository.
Moreover, the content of the opal-gui folder has been moved to the
root of the repository.
Finally, I added the Netbeans IDE file "nbactions.xml" to the .hgignore
file.