lib: Update license information on java, properties, xsd and xml files
All java, properties, xsd and xml files in the library component have
now a proper license: CeCILL-C.

For easier maintenance, the maven-license-plugin maven plugin was
used. It is able to read the content of the LICENSE.txt file and
update source files content. Thus, if there is any modification of
the license to perform, you juste have to update the LICENSE.txt
file and use the maven command "mvn license:format" to update
automatically all source files.