CardConfigFactory: Reformat code and fix a unit test
Some minor comments were translated from french to english.

Moreover, an issue about the key type was fixed in the CardConfigFactory
class in the revision 896706843812 but unit tests were not updated.
The current revision also updates the corresponding unit test
"testSaveLocalCardConfigsToXML" by removing the
SCGPKey using the key type MOTHER_KEY in the test and adding SCGemVisa and
SCGemVisa2 keys.

Minor modifications in HttpPostResponseTest and RamOverHttpTest are tied to the
presence of invisible indentations (corrected thanks to my IDE).